chore(converter): move get_all_currency_codes func to util/currency/converter.rs
