+++ src/handlers/markups/mod.rs
@@ -1,3 +1,2 @@
-pub mod currency_keyboard;
 pub mod inlines;
 pub mod markuper;
