+++ src/handlers/messages/chat.rs
@@ -1,6 +1,7 @@
 use crate::db::schemas::group::Group;
-use crate::handlers::markups::currency_keyboard::get_all_currency_codes;
-use crate::util::currency::converter::{CURRENCY_CONFIG_PATH, CurrencyStruct};
+use crate::util::currency::converter::{
+    CURRENCY_CONFIG_PATH, CurrencyStruct, get_all_currency_codes,
+};
 use crate::util::errors::MyError;
 use log::{error, info};
 use oximod::ModelTrait;
@@ -12,7 +13,7 @@ use teloxide::types::{ChatMemberUpdated, ParseMode};
 pub async fn handle_bot_added(bot: Bot, update: ChatMemberUpdated) -> Result<(), MyError> {
     let id = update.chat.id.to_string();
     info!("New chat added. ID: {}", id);
-    
+
     let all_codes = get_all_currency_codes(CURRENCY_CONFIG_PATH.parse().unwrap())?;
 
     let necessary_codes = all_codes
